week_6/requirements.txt
iniconfig==2.1.0
mirakuru==2.6.0
numpy==2.2.4
packaging==25.0
pandas==2.2.3
pluggy==1.6.0
port-for==0.7.4
psutil==7.0.0
psycopg==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
pytest==7.4.4
pytest-postgresql==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==0.15.0
pytz==2025.1
requests==2.32.3
setuptools==75.8.2
six==1.17.0
tzdata==2025.1
